# schemas/generated/openapi.json
{"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"}}},"info":{"title":"DRL CO Experiment Platform","version":"1.0.0"},"openapi":"3.1.0","paths":{"/api/v1/analyses":{"get":{"operationId":"analyses_api_v1_analyses_get","parameters":[{"in":"query","name":"family","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Family"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Analyses"},"post":{"operationId":"analysis_create_api_v1_analyses_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Analysis Create"}},"/api/v1/analyses/{aid}":{"get":{"operationId":"analysis_api_v1_analyses__aid__get","parameters":[{"in":"path","name":"aid","required":true,"schema":{"title":"Aid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Analysis"}},"/api/v1/analyses/{aid}/artifacts/{name}":{"get":{"operationId":"analysis_artifact_api_v1_analyses__aid__artifacts__name__get","parameters":[{"in":"path","name":"aid","required":true,"schema":{"title":"Aid","type":"string"}},{"in":"path","name":"name","required":true,"schema":{"title":"Name","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Analysis Artifact"}},"/api/v1/batches":{"post":{"operationId":"submit_api_v1_batches_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"202":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Submit"}},"/api/v1/batches/{bid}":{"get":{"operationId":"batch_api_v1_batches__bid__get","parameters":[{"in":"path","name":"bid","required":true,"schema":{"title":"Bid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Batch"}},"/api/v1/capabilities":{"get":{"operationId":"capabilities_api_v1_capabilities_get","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Capabilities"}},"/api/v1/comparisons":{"post":{"operationId":"compare_api_v1_comparisons_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Compare"}},"/api/v1/datasets":{"get":{"operationId":"datasets_api_v1_datasets_get","parameters":[{"in":"query","name":"family","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Family"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Datasets"},"post":{"operationId":"dataset_create_api_v1_datasets_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Dataset Create"}},"/api/v1/datasets/freeze":{"post":{"operationId":"dataset_freeze_api_v1_datasets_freeze_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Dataset Freeze"}},"/api/v1/datasets/generate":{"post":{"operationId":"dataset_generate_api_v1_datasets_generate_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Dataset Generate"}},"/api/v1/datasets/import":{"post":{"operationId":"dataset_import_api_v1_datasets_import_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Dataset Import"}},"/api/v1/datasets/validate":{"post":{"operationId":"dataset_validate_api_v1_datasets_validate_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Dataset Validate"}},"/api/v1/datasets/{did}/archive":{"post":{"operationId":"dataset_archive_api_v1_datasets__did__archive_post","parameters":[{"in":"path","name":"did","required":true,"schema":{"title":"Did","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Dataset Archive"}},"/api/v1/datasets/{did}/revisions/{revision}":{"get":{"operationId":"dataset_revision_api_v1_datasets__did__revisions__revision__get","parameters":[{"in":"path","name":"did","required":true,"schema":{"title":"Did","type":"string"}},{"in":"path","name":"revision","required":true,"schema":{"title":"Revision","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Dataset Revision"}},"/api/v1/debug/step":{"post":{"operationId":"debug_step_api_v1_debug_step_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Debug Step"}},"/api/v1/diagnostics":{"post":{"operationId":"diagnostics_api_v1_diagnostics_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Diagnostics"}},"/api/v1/experiments/plan":{"post":{"operationId":"experiment_plan_api_v1_experiments_plan_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Experiment Plan"}},"/api/v1/experiments/validate":{"post":{"operationId":"validate_api_v1_experiments_validate_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Validate"}},"/api/v1/families":{"get":{"operationId":"families_api_v1_families_get","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Families"}},"/api/v1/health":{"get":{"operationId":"health_api_v1_health_get","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Health"}},"/api/v1/metrics":{"get":{"operationId":"metrics_api_v1_metrics_get","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Metrics"}},"/api/v1/plugins":{"get":{"operationId":"plugins_api_v1_plugins_get","parameters":[{"in":"query","name":"family","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Family"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Plugins"}},"/api/v1/reproduction/export":{"post":{"operationId":"export_api_v1_reproduction_export_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Export"}},"/api/v1/reproduction/validate":{"post":{"operationId":"validate_reproduction_api_v1_reproduction_validate_post","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Validate Reproduction"}},"/api/v1/results/export":{"post":{"operationId":"results_export_api_v1_results_export_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Config","type":"object"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Results Export"}},"/api/v1/runs":{"get":{"operationId":"runs_api_v1_runs_get","parameters":[{"in":"query","name":"family","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Family"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Runs"}},"/api/v1/runs/{rid}":{"get":{"operationId":"run_api_v1_runs__rid__get","parameters":[{"in":"path","name":"rid","required":true,"schema":{"title":"Rid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Run"}},"/api/v1/runs/{rid}/artifacts/{artifact_id}":{"get":{"operationId":"artifact_api_v1_runs__rid__artifacts__artifact_id__get","parameters":[{"in":"path","name":"rid","required":true,"schema":{"title":"Rid","type":"string"}},{"in":"path","name":"artifact_id","required":true,"schema":{"title":"Artifact Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Artifact"}},"/api/v1/runs/{rid}/cancel":{"post":{"operationId":"cancel_api_v1_runs__rid__cancel_post","parameters":[{"in":"path","name":"rid","required":true,"schema":{"title":"Rid","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Cancel"}},"/api/v1/runs/{rid}/events":{"get":{"operationId":"event_list_api_v1_runs__rid__events_get","parameters":[{"in":"path","name":"rid","required":true,"schema":{"title":"Rid","type":"string"}},{"in":"query","name":"after","required":false,"schema":{"default":0,"title":"After","type":"integer"}},{"in":"query","name":"stream","required":false,"schema":{"default":false,"title":"Stream","type":"boolean"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Event List"}},"/api/v1/runs/{rid}/rerun":{"post":{"operationId":"rerun_api_v1_runs__rid__rerun_post","parameters":[{"in":"path","name":"rid","required":true,"schema":{"title":"Rid","type":"string"}}],"responses":{"202":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Rerun"}},"/api/v1/schemas/{schema_id}":{"get":{"operationId":"schema_api_v1_schemas__schema_id__get","parameters":[{"in":"path","name":"schema_id","required":true,"schema":{"title":"Schema Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Schema"}},"/api/v1/session":{"get":{"operationId":"session_api_v1_session_get","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Session"}},"/api/v1/shutdown":{"post":{"operationId":"shutdown_api_v1_shutdown_post","responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"}},"summary":"Shutdown"}}}}